+++ lib/frp/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name    = "enso-frp"
+version = "0.1.0"
+authors = ["Enso Team <contact@luna-lang.org>"]
+edition = "2018"
+
+[lib]
+
+[dependencies]
+basegl-prelude             = { version = "0.1.0"  , path = "../prelude"       }
+basegl-system-web          = { version = "0.1.0"  , path = "../system/web"    }
+percent-encoding           = { version = "2.1.0"  }
